change the fitter code (fix the difference between data and mc)
